+++ editors/vscode/standalone-client-state.js
@@ -1,0 +1,17 @@
+"use strict";
+
+function transitionStandaloneClient(project, running) {
+	project.clientRunning = running;
+	if (!running) {
+		project.forwardedDocuments.clear();
+		return false;
+	}
+	if (!project.clientStarted) {
+		project.clientStarted = true;
+		return false;
+	}
+	project.forwardedDocuments.clear();
+	return true;
+}
+
+module.exports = { transitionStandaloneClient };